Adds support for OFT to EML
OFT is very similar to MSG, I realized simply renaming files from .oft to .msg allowed MsgViewer to convert it to .eml with no issue.

Instead of needing to rename the source files each time, it would be nice to have the option within the application natively. This seems to work, thanks!